+++ point_cloud_transport/CMakeLists.txt
@@ -18,8 +18,9 @@ find_package(message_filters REQUIRED)
 find_package(pluginlib REQUIRED)
 find_package(rclcpp REQUIRED)
 find_package(rclcpp_components REQUIRED)
-find_package(sensor_msgs REQUIRED)
 find_package(rcpputils REQUIRED)
+find_package(rmw REQUIRED)
+find_package(sensor_msgs REQUIRED)
 
 # Build libpoint_cloud_transport
 add_library(${PROJECT_NAME}
@@ -40,12 +41,12 @@ target_include_directories(${PROJECT_NAME} PUBLIC
   "$<INSTALL_INTERFACE:include/${PROJECT_NAME}>")
 target_link_libraries(${PROJECT_NAME} PUBLIC
   message_filters::message_filters
+  pluginlib::pluginlib
   rclcpp::rclcpp
+  rcpputils::rcpputils
+  rmw::rmw
   ${sensor_msgs_TARGETS}
-  rcpputils::rcpputils)
-target_link_libraries(${PROJECT_NAME} PRIVATE
-  pluginlib::pluginlib)
-
+)
 target_compile_definitions(${PROJECT_NAME} PRIVATE "POINT_CLOUD_TRANSPORT_BUILDING_DLL")
 
 # Build libpoint_cloud_transport_plugins
@@ -53,10 +54,11 @@ add_library(${PROJECT_NAME}_plugins
   src/manifest.cpp
 )
 add_library(${PROJECT_NAME}::${PROJECT_NAME}_plugins ALIAS ${PROJECT_NAME}_plugins)
-
 target_link_libraries(${PROJECT_NAME}_plugins PRIVATE
   ${PROJECT_NAME}
-  pluginlib::pluginlib)
+  pluginlib::pluginlib
+  ${sensor_msgs_TARGETS}
+)
 
 add_library(pc_republish_node SHARED src/republish.cpp)
 target_link_libraries(pc_republish_node PRIVATE
@@ -106,7 +108,7 @@ ament_export_include_directories("include/${PROJECT_NAME}")
 
 ament_export_targets(export_${PROJECT_NAME})
 
-ament_export_dependencies(message_filters rclcpp sensor_msgs pluginlib rcpputils)
+ament_export_dependencies(message_filters rclcpp sensor_msgs pluginlib rcpputils rmw)
 
 if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
@@ -121,67 +123,43 @@ if(BUILD_TESTING)
 
   ament_add_gtest(${PROJECT_NAME}-publisher test/test_publisher.cpp)
   if(TARGET ${PROJECT_NAME}-publisher)
-    target_link_libraries(${PROJECT_NAME}-publisher ${PROJECT_NAME})
-    ament_target_dependencies(${PROJECT_NAME}-publisher
-      pluginlib
-    )
+    target_link_libraries(${PROJECT_NAME}-publisher ${PROJECT_NAME} rclcpp::rclcpp)
   endif()
 
   ament_add_gtest(${PROJECT_NAME}-codec test/test_point_cloud_codec.cpp)
   if(TARGET ${PROJECT_NAME}-codec)
-    target_link_libraries(${PROJECT_NAME}-codec ${PROJECT_NAME})
-    ament_target_dependencies(${PROJECT_NAME}-codec
-      pluginlib
-    )
+    target_link_libraries(${PROJECT_NAME}-codec ${PROJECT_NAME} rclcpp::rclcpp ${sensor_msgs_TARGETS})
   endif()
 
   ament_add_gtest(${PROJECT_NAME}-subscriber test/test_subscriber.cpp)
   if(TARGET ${PROJECT_NAME}-subscriber)
-    target_link_libraries(${PROJECT_NAME}-subscriber ${PROJECT_NAME})
-    ament_target_dependencies(${PROJECT_NAME}-subscriber
-      pluginlib
-    )
+    target_link_libraries(${PROJECT_NAME}-subscriber ${PROJECT_NAME} rclcpp::rclcpp ${sensor_msgs_TARGETS})
   endif()
 
   ament_add_gtest(${PROJECT_NAME}-message_passing test/test_message_passing.cpp)
   if(TARGET ${PROJECT_NAME}-message_passing)
-    target_link_libraries(${PROJECT_NAME}-message_passing ${PROJECT_NAME})
-    ament_target_dependencies(${PROJECT_NAME}-message_passing
-      pluginlib
-    )
+    target_link_libraries(${PROJECT_NAME}-message_passing ${PROJECT_NAME} rclcpp::rclcpp ${sensor_msgs_TARGETS})
   endif()
 
   ament_add_gtest(${PROJECT_NAME}-qos_override test/test_qos_override.cpp)
   if(TARGET ${PROJECT_NAME}-qos_override)
-    target_link_libraries(${PROJECT_NAME}-qos_override ${PROJECT_NAME})
-    ament_target_dependencies(${PROJECT_NAME}-qos_override
-      pluginlib
-    )
+    target_link_libraries(${PROJECT_NAME}-qos_override ${PROJECT_NAME} rclcpp::rclcpp ${sensor_msgs_TARGETS})
   endif()
 
   ament_add_gtest(${PROJECT_NAME}-remapping test/test_remapping.cpp)
   if(TARGET ${PROJECT_NAME}-remapping)
-    target_link_libraries(${PROJECT_NAME}-remapping ${PROJECT_NAME})
-    ament_target_dependencies(${PROJECT_NAME}-remapping
-      pluginlib
-    )
+    target_link_libraries(${PROJECT_NAME}-remapping ${PROJECT_NAME} rclcpp::rclcpp ${sensor_msgs_TARGETS})
   endif()
 
   ament_add_gtest(${PROJECT_NAME}-message-filter test/test_message_filter.cpp)
   if(TARGET ${PROJECT_NAME}-message-filter)
-    target_link_libraries(${PROJECT_NAME}-message-filter ${PROJECT_NAME})
-    ament_target_dependencies(${PROJECT_NAME}-message-filter
-      message_filters
-      pluginlib
-    )
+    target_link_libraries(${PROJECT_NAME}-message-filter
+      ${PROJECT_NAME} message_filters::message_filters rclcpp::rclcpp ${sensor_msgs_TARGETS})
   endif()
 
   ament_add_gtest(${PROJECT_NAME}-single_subscriber_publisher test/test_single_subscriber_publisher.cpp)
   if(TARGET ${PROJECT_NAME}-single_subscriber_publisher)
-    target_link_libraries(${PROJECT_NAME}-single_subscriber_publisher ${PROJECT_NAME})
-    ament_target_dependencies(${PROJECT_NAME}-single_subscriber_publisher
-      pluginlib
-    )
+    target_link_libraries(${PROJECT_NAME}-single_subscriber_publisher ${PROJECT_NAME} rclcpp::rclcpp ${sensor_msgs_TARGETS})
   endif()
 endif()
