+++ point_cloud_transport/include/point_cloud_transport/publisher.hpp
@@ -32,8 +32,8 @@
 #ifndef POINT_CLOUD_TRANSPORT__PUBLISHER_HPP_
 #define POINT_CLOUD_TRANSPORT__PUBLISHER_HPP_
 
-#include <string>
 #include <memory>
+#include <string>
 
 #include "rclcpp/macros.hpp"
 #include "rclcpp/node.hpp"
